软件工程/软件架构与中间件/实验/SA_lab3/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
DBUtils==3.1.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.7.0
eventlet==0.37.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
Flask-SocketIO==5.4.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langgraph==0.2.38
langgraph-checkpoint==2.0.1
langgraph-sdk==0.1.33
langsmith==0.1.129
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.4
nest_asyncio==1.6.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.5.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.1.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data==0.6.2
streamlit==1.38.0
tenacity==8.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
waitress==2.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
